refactor: move SubjectIconBrushBorders to TeacherComponents
