+++ src/components/TeacherComponents/LessonOverviewHeaderMobile/LessonOverviewHeaderMobile.tsx
@@ -6,7 +6,7 @@ import { LessonOverviewHeaderShareAllButton } from "@/components/TeacherComponen
 import Box from "@/components/SharedComponents/Box";
 import Flex from "@/components/SharedComponents/Flex";
 import { Heading, P, Span } from "@/components/SharedComponents/Typography";
-import SubjectIconBrushBorders from "@/components/SubjectIconBrushBorders";
+import SubjectIconBrushBorders from "@/components/TeacherComponents/SubjectIconBrushBorders";
 import LessonMetadata from "@/components/SharedComponents/LessonMetadata";
 
 export const LessonOverviewHeaderMobile: FC<LessonOverviewHeaderProps> = (
